Move app to components
